chatbot/requirements.txt
flask==3.1.*
gunicorn==23.*
google-genai==1.*
google-auth==2.*
